Use correct size for curlim and maxlim.

+++ src/limits.c
@@ -352,8 +352,8 @@ serialize_limits(char **info, size_t info_max)
     for (idx = 0; idx < nitems(saved_limits); idx++) {
 	const struct saved_limit *lim = &saved_limits[idx];
 	const struct rlimit *rl = &lim->oldlimit;
-	char curlim[(((sizeof(int) * 8) + 2) / 3) + 2];
-	char maxlim[(((sizeof(int) * 8) + 2) / 3) + 2];
+	char curlim[(((sizeof(long long) * 8) + 2) / 3) + 2];
+	char maxlim[(((sizeof(long long) * 8) + 2) / 3) + 2];
 
 	if (!lim->saved)
 	    continue;
